extract AnnotationTargetResolver
